+++ en/notifications.js
@@ -1,0 +1,4 @@
+// Notifications & flashnews
+export default {
+  'flashnews.create': 'Create a new news flash'
+}